feat(workout): add workout by types component with infinite scroll and loading state
