+++ libs/ui/src/index.ts
@@ -34,6 +34,7 @@ export * from './lib/components/detail-creator-profile/detail-creator-profile.co
 export * from './lib/components/card-workout-info/card-workout-info.component';
 export * from './lib/components/card-with-lateral-image/card-with-lateral-image.component';
 export * from './lib/components/modal-add-payment-method/modal-add-payment-method.component';
+export * from './lib/components/workout-by-types/workout-by-types.component';
 
 export * from './lib/components/payment-frequency-settings/payment-frequency-settings.component';
 export * from './lib/components/reset-password-form/reset-password-form/reset-password-form.component';
