+++ libs/ui/src/lib/components/detail-creator-profile/detail-creator-profile.component.scss
@@ -1,11 +1,16 @@
 :host {
   display: block;
-
+  .goback-button {
+    position: absolute;
+    top: var(--ion-safe-area-top);
+    left: 16px;
+    z-index: 10;
+  }
   .detail-creator-profile-content {
     ::ng-deep {
       .detail-creator-profile-page {
         margin-top: -30px !important ;
-        margin-bottom: 40px !important;
+        // margin-bottom: 40px !important;
         min-height: auto !important;
       }
