+++ libs/ui/src/lib/components/workout-by-types/workout-by-types.component.html
@@ -1,0 +1,11 @@
+<div class="container-workout-list" style="margin-bottom: 50px">
+  @if (!isLoadingWorkouts()) { @for (workout of workouts(); track
+  workout.workoutId) {
+  <lib-card-workout-info [workout]="workout"></lib-card-workout-info>
+  } } @else {
+  <app-list-skeleton [itemRepeat]="3"></app-list-skeleton>
+  }
+</div>
+<ion-infinite-scroll (ionInfinite)="onIonInfinite($event)">
+  <ion-infinite-scroll-content></ion-infinite-scroll-content>
+</ion-infinite-scroll>
