+++ libs/ui/src/lib/components/workout-by-types/workout-by-types.component.scss
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
